[Bronze V] Title: 이상한 기호, Time: 68 ms, Memory: 11500 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -6,9 +6,9 @@ public class Main {
 	public static void main(String[] args) throws Exception{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int A = Integer.parseInt(st.nextToken());
-		int B = Integer.parseInt(st.nextToken());
-		System.out.print((A+B)*(A-B));
+		long A = Long.parseLong(st.nextToken());
+		long B = Long.parseLong(st.nextToken());
+		System.out.print((long)((A+B)*(A-B)));
 	}
 
 }
